CORS configuration moved from Main class to SecurityConfig
